[#217] Move parsing specs to separate modules
